add header avatar config
