+++ src/components/templates/Header/Header.tsx
@@ -1,5 +1,5 @@
 import {
-  IconButton, Menu, Typography, Tooltip,
+  IconButton,
 } from '@mui/material';
 import { useState } from 'react';
 import { useLocation } from 'react-router-dom';
@@ -15,12 +15,13 @@ import { AppIcon } from '../../UI/Icons';
 import { BrandLogoName } from '../BrandLogoName';
 import { GuestUserModal, LoggedUserDrawer, NotLoggedDrawer } from './features';
 import {
-  AnchorButton, AppColors, FlexContainer, PrimaryButton, SecondaryButton, Avatar, MenuItem,
+  AnchorButton, AppColors, FlexContainer, PrimaryButton, SecondaryButton,
 } from '../../../styles';
 import {
   GuestUserButton,
   HeaderContainer, HeaderNav, HeaderNavAnchor, HeaderShadow,
 } from './Header.styled';
+import { HeaderAvatarConfig } from '../HeaderAvatarConfig';
 
 const Header = ({ isLandingPage = false }: HeaderProps) => {
   const location = useLocation();
@@ -30,7 +31,6 @@ const Header = ({ isLandingPage = false }: HeaderProps) => {
   const isMobile = windowSize === 'Mobile';
   const isDesktop = windowSize === 'Desktop';
 
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const [openNotLoggedDrawer, setOpenNotLoggedDrawer] = useState(false);
   const [openLoggedDrawer, setOpenLoggedDrawer] = useState(false);
   const [openGuestUserModal, setOpenGuestUserModal] = useState(false);
@@ -42,12 +42,6 @@ const Header = ({ isLandingPage = false }: HeaderProps) => {
   const handleGuestUserModalMobile = () => {
     toggleNotLoggedDrawer();
     toggleGuestUserModal();
-  };
-  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
   };
 
   const activeDashboardPage = location.pathname === DASHBOARD_ROUTE;
@@ -65,34 +59,7 @@ const Header = ({ isLandingPage = false }: HeaderProps) => {
             </HeaderNav>
           ) }
           { (!isGuestUser && userLoggedOn && isDesktop) && (
-            <>
-              <Tooltip title={
-                <Typography>Abrir configuración</Typography>
-              }
-              >
-                <IconButton aria-label="open-configuration-button" onClick={handleMenu}>
-                  <Avatar>{initials}</Avatar>
-                </IconButton>
-              </Tooltip>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: 'bottom',
-                  horizontal: 'right',
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                open={Boolean(anchorEl)}
-                onClose={handleClose}
-              >
-                <MenuItem onClick={handleClose}>Profile</MenuItem>
-                <MenuItem onClick={handleClose}>My account</MenuItem>
-              </Menu>
-            </>
+            <HeaderAvatarConfig initials={initials} />
           ) }
           {/** TODO: Change this to use avatar */}
           { (!isGuestUser && !userLoggedOn && !isMobile) && (
