+++ src/components/templates/HeaderAvatarConfig/index.ts
@@ -1,0 +1,1 @@
+export { HeaderAvatarConfig } from './HeaderAvatarConfig';
